Trying out an alternate mapping for paragraphs.
By default vim's paragraph text object motions consider a paragraph to
be one or more completely empty lines (i.e. no whitespace) between
blocks of text.

I think I'd prefer to be able to jump to the first non-whitespace
character of the next block instead. And I'd like lines to be treated as
empty even if the contain some whitespace.

One caveat is that in visual mode we should jump to the last line of the
block, instead of the first, to allow for neat selection of blocks.

I also added a mapping to easily start a new block with your cursor on
the first line of an existing one.